allow floats for fps

+++ jsk_rosbag_tools/scripts/video_to_bag.py
@@ -16,7 +16,7 @@ def main():
     parser.add_argument('--topic-name', type=str,
                         default='/video/rgb/image_raw',
                         help='Converted topic name.')
-    parser.add_argument('--fps', type=int,
+    parser.add_argument('--fps', type=float,
                         help='Frame Rate.', default=None)
     parser.add_argument('--compress', action='store_true',
                         help='Compress Image flag.')
